Modified tests for StatsCommand"
